Use attachment timestamp in filename of exported screenshots
This allows consumers of the screenshots to easily sort the screenshots in the
order that they were taken.

Note: The modification date of the exported screenshot files can be used
for the same purpose in most scenarios. However, the precision of the
modification date isn't kept intact when the screenshots are zipped and
then unzipped. In that scenario it's useful for the screenshots to have
the timestamp in their name as well.